+++ NEWS.md
@@ -80,6 +80,7 @@
 * Data Viewer: fixed a crash that could occur when calling 'View()' on matrix
 * R Markdown: hide chunk toolbar when chunk is hidden by folded Markdown section
 * R Markdown: fixed highlighting with multiple Markdown links on one line
+* R Markdown: improved highlighting for '\*', '\_' text
 * R Markdown: fixed issue where sections did not display in document outline when following empty bulleted list
 * Reformat code: fixed an issue where '\*\*' was not recognized as a single token
 * SVN: avoid recursive revert when reverting changes to modified directory properties